Checkpoint: Correction complète de l'accès spectateur et synchronisation du curseur avec coordonnées relatives :
**Problèmes corrigés :**

1. **Accès spectateur sans authentification**
   - ✅ Désactivation de la redirection automatique vers login pour les pages `/view` et `/viewer`
   - ✅ Les spectateurs peuvent accéder directement à la présentation avec uniquement le code de session
   - ✅ Toutes les procédures tRPC utilisées par les spectateurs sont déjà en `publicProcedure`
   - ✅ Modification de `main.tsx` pour exclure les pages spectateur de la redirection auth

2. **Curseur synchronisé avec coordonnées relatives**
   - ✅ Modification du schéma : `cursorX` et `cursorY` passent de `int` (pixels) à `float` (pourcentage 0-100)
   - ✅ Migration de base de données appliquée avec succès
   - ✅ Le présentateur envoie les coordonnées en **pourcentage** relatif à la taille de l'image
   - ✅ Le spectateur affiche le curseur rouge avec ces pourcentages
   - ✅ Le curseur s'affiche correctement **quel que soit la taille de l'écran**

3. **Interface spectateur simplifiée**
   - ✅ **Pas de barre de zoom** côté spectateur (supprimée)
   - ✅ **Curseur rouge visible** pour les spectateurs quand zoom > 100%
   - ✅ Zoom du présentateur appliqué automatiquement
   - ✅ Interface épurée pour une meilleure expérience de visualisation

**Modifications techniques :**

**Schéma de base de données (drizzle/schema.ts) :**
```typescript
export const presenterCursors = mysqlTable("presenter_cursors", {
  // ...
  /** Cursor X position as percentage (0-100) relative to the document width */
  cursorX: float("cursorX").default(0).notNull(),
  /** Cursor Y position as percentage (0-100) relative to the document height */
  cursorY: float("cursorY").default(0).notNull(),
  // ...
});
```

**Présentateur (client/src/pages/PresenterControl.tsx) :**
```typescript
const handleMouseMove = (e: React.MouseEvent<HTMLDivElement>) => {
  const rect = e.currentTarget.getBoundingClientRect();
  const x = e.clientX - rect.left;
  const y = e.clientY - rect.top;

  // Convert pixel coordinates to percentage (0-100)
  const xPercent = (x / rect.width) * 100;
  const yPercent = (y / rect.height) * 100;

  updateZoomAndCursorMutation.mutate({
    sessionId: sessionIdNum,
    zoomLevel: zoom,
    cursorX: xPercent,  // Percentage instead of pixels
    cursorY: yPercent,  // Percentage instead of pixels
    cursorVisible: showMouseCursor && zoom > 100,
  });
};
```

**Spectateur (client/src/pages/Viewer.tsx) :**
```typescript
{/* Presenter cursor - Red, visible when zoom > 100% */}
{presenterCursorVisible && presenterZoom > 100 && (
  <div
    className="absolute w-6 h-6 border-2 border-red-500 rounded-full"
    style={{
      left: `${presenterCursorX}%`,   // Percentage positioning
      top: `${presenterCursorY}%`,    // Percentage positioning
      transform: "translate(-50%, -50%)",
    }}
  >
    <div className="absolute inset-1 border-2 border-red-500 rounded-full opacity-50" />
  </div>
)}
```

**Authentification (client/src/main.tsx) :**
```typescript
const redirectToLoginIfUnauthorized = (error: unknown) => {
  // ...
  // Don't redirect if user is on viewer page (public access)
  const isViewerPage = window.location.pathname.startsWith('/view') ||
                       window.location.pathname === '/viewer';
  if (isViewerPage) return;

  window.location.href = getLoginUrl();
};
```

**Avantages de l'approche par pourcentage :**
- ✅ **Indépendant de la résolution** : Le curseur s'affiche au bon endroit quel que soit la taille de l'écran
- ✅ **Responsive** : Fonctionne sur mobile, tablette et desktop
- ✅ **Précis** : Les coordonnées relatives garantissent une position exacte
- ✅ **Scalable** : Supporte tous les ratios d'aspect d'image

**Fonctionnalités conservées du checkpoint précédent :**
- ✅ Upload d'images depuis mobile (Data URL base64)
- ✅ Bouton "Ajouter Document" amélioré pour mobile
- ✅ Support caméra mobile (`capture="environment"`)
- ✅ Barre de zoom (slider 50-200%) côté présentateur
- ✅ Déplacement du document (pan) avec souris et doigts
- ✅ Zoom tactile (pinch-to-zoom)
- ✅ Synchronisation du zoom en temps réel
- ✅ Partage WhatsApp
- ✅ Compteur de spectateurs
- ✅ Format portrait/paysage

+++ drizzle/schema.ts
@@ -1,4 +1,4 @@
-import { int, mysqlEnum, mysqlTable, text, timestamp, varchar, boolean } from "drizzle-orm/mysql-core";
+import { int, mysqlEnum, mysqlTable, text, timestamp, varchar, boolean, float } from "drizzle-orm/mysql-core";
 
 /**
  * Core user table backing auth flow.
@@ -137,10 +137,10 @@ export const presenterCursors = mysqlTable("presenter_cursors", {
   sessionId: int("sessionId").notNull(),
   /** Current zoom level (50-200) */
   zoomLevel: int("zoomLevel").default(100).notNull(),
-  /** Cursor X position (relative to the document) */
-  cursorX: int("cursorX").default(0).notNull(),
-  /** Cursor Y position (relative to the document) */
-  cursorY: int("cursorY").default(0).notNull(),
+  /** Cursor X position as percentage (0-100) relative to the document width */
+  cursorX: float("cursorX").default(0).notNull(),
+  /** Cursor Y position as percentage (0-100) relative to the document height */
+  cursorY: float("cursorY").default(0).notNull(),
   /** Whether the cursor should be visible */
   cursorVisible: boolean("cursorVisible").default(false).notNull(),
   /** Timestamps */
